memory: MapError 可见性与强类型收口 + dead API 收敛
- MapError 提升为 pub 并经 pub use re-export 到 memory 根（pub API 签名可命名）
- map_device 改收 PhysAddr（与 Device.base 对齐），移除 usize→PhysAddr 往返
- root_page() 返回类型 u64→usize，与 switch_space/satp::make 统一
- 删除 dead API：VirtAddr::new/is_kernel、PageTableEntry empty/new/new_branch/is_user/as_u64
- 移除 VirtAddr::page_align 误标 #[allow(dead_code)]（fault.rs 在用）
- VA/PA 收敛式对称：成对方法两边保留，单边 dead 删除；mmap 系 4 方法保留为预留

+++ src/init.rs
@@ -73,7 +73,7 @@ pub unsafe fn run() -> Result<()> {
         .root_page();
     info!(
         "address space ready (Sv39 root table {:#x})",
-        (root_ppn << crate::memory::PAGE_SHIFT) as usize
+        root_ppn << crate::memory::PAGE_SHIFT
     );
 
     // trap 必须先于 driver：设备探测会访问 DTB/MMIO，若期间异常而 stvec 未就位，
